requirements.txt
alembic==1.4.1
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
blis==0.7.5
Boruta==0.3
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
databricks-cli==0.16.2
dataclasses==0.8
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.10
djangorestframework==3.13.1
docker==5.0.3
entrypoints==0.3
Flask==2.0.2
funcy==1.17
future==0.18.2
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.20
greenlet==1.1.2
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.7.0
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.0.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
kmodes==0.11.1
lightgbm==3.3.1
llvmlite==0.36.0
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.3.4
missingno==0.5.0
mistune==0.8.4
mlflow==1.22.0
mlxtend==0.19.0
multimethod==1.5
murmurhash==1.0.6
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.5.1
nltk==3.6.7
notebook==6.4.6
numba==0.53.1
numexpr==2.8.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.7.1
patsy==0.5.2
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==8.4.0
plac==1.1.3
plotly==5.5.0
preshed==3.0.6
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.7
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pycaret==2.3.5
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.1
pyLDAvis==3.3.1
pynndescent==0.5.5
pyod==0.9.6
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
regex==2021.11.10
requests==2.26.0
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==2.3.7
SQLAlchemy==1.4.29
sqlparse==0.4.2
srsly==1.0.5
statsmodels==0.12.2
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
thinc==7.4.5
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==4.0.1
umap-learn==0.5.2
urllib3==1.26.7
visions==0.7.4
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wordcloud==1.8.1
yellowbrick==1.3.post1
zipp==3.6.0
